Added data file cutting and splicing.
[Details]
For certain tests in BugsInPy, the test refers to a data file that has
been added in that version accompanying the test. This is obviously
needed in order to run the test, and hence the cut and splice scripts
have been ammended to find and copy over these data files when needed.

[Note]
The cut script grabs the data file names rudimentarily from the
constants used in the test method. This of course does not always work,
and that is a concern. To make this more robust, it searches for any
data file with the same name, regardless of the file extension (as it is
often the case that the extension is added in another method).

[File's changed]

	modified:   cut
	modified:   splice