Added handling of multiline imports.
[Details]
Added proper handling of multiline imports. The cut function now
retrieves the exact multiline import that the required module is in, and
adds this to the blocks. The splice method then takes these imports and
searches for similar imports in the new version, and overrides them if
found, if not found, it will add the new import. The splice method will
adapt imports that are already there but contain different sets of
packages. In particular, it will only add new packages that are
necessary for the test method. This means it does not add the full
import given by the cut method.

[Note]
TODO: refactor the import merging logic to the cut method, so that only
packages that are necessary in the full import are copied over.

[File's changed]

	modified:   cut
	modified:   splice