Improved error handling for module not found.
[Details]
Improved the handling of errors caused by modules not being found in
the cut and splice scripts. This does not alleviate the erros, just
allows for better debugging.

[Note]
Importing the test module in the cut and splice scripts has proved to
cause quite a few problems in different projects when either the test
does not exist, or imports something that is not in the cut or splice
scripts path. Should try to fix this.

[File's changed]

	modified:   bugsinpy-cut
	modified:   bugsinpy-splice