Fixed the transplanting of functions in cut.
[Details]
Fixed some bugs and logic in the function transplanting part of cut.
This included allowing the CALL_FUNCTION_KW bytecode command for keyword
arguments to be processed, and also allowing all imports to be treated
as functions first, to see if they exist in the code.

[Bugs]
Fixed minor bug with conditional checking.

[File's changed]

	modified:   framework/bin/bugsinpy-cut