Fixes for bugsinpy-cut.
[Details]
The bugsinpy-cut script works by importing the test cases as if they
were modules. This however does not always work nicely as they are not
in fact modules, and python usually has a library called 'tests' which
is imported instead of the usual 'tests' directory. To fix this, the
bugsinpy-cut script now temporarily adds an __init__.py file to the
tests directory to force it to be a module for the duration of the
script.

[Note]
Also added catch for ImportError's

[File's changed]

	modified:   bugsinpy-cut