Fixing bugs in BugsInPy.
[Details]
There were a few issues with the bugsinpy-compile script which were
causing problems with the multifaulting. These needed to be fixed by
changing the bugsinpy-compile script.

[Bugs]
1. The bugsinpy-compile script allowed certain python versions to expose
certain libraries by adding them to the global .bash_rc. Although this
works for that particular version, when this was done for multiple
versions the .bash_rc then contained all of them, and python would then
get confused as to which version to use. To remedy this, these libraries
were exposed only locally for the project by adding them to a .pth file
in the site-packages directory for python. This allowed the same
functionality on a local level.

2. The BugsInPy dataset contained a python version for every bug
version, contained in that versions bugsinpy-bug.info file, however this
information was not used. In order for all the bug versions to run
properly, the bugsinpy-compile script was modified to use this version
by locally installing the python version it uses and then using this for
the virtual environment for that bug version. The python versions are
now stored in the directory python_versions inside the BugsInPy source
directory, which allows reuse between bug versions, and does not change
the user's system defaults for python (or storage on /usr/bin).

[File's changed]

	modified:   bugsinpy-compile