Changed version search to use LCS algorithm.
[Details]
The comparison of the error messages in version_search.sh was done on
full equality, which is the only comparison available in bash. Many of
the errors, however, only differ in small ordering differences, or
changes to minor names or numbers. This made it difficult to detect that
these bugs were actually exposed in the previous version. To remedy
this, the longest common subsequence algorithm was implemented in java
and used in the version search to check for error messages that are
mostly similar. A threshold value of 85% of the error message being the
same is set to verify error messages that are deemed the same.

[File's changed]

	new file:   lcs.class
	new file:   lcs.java
	modified:   version_search.sh