Allow function unwrapping and fixed class line.
[Details]
Changed the implementation for function unwrapping to be allowed always,
instead of only when the function was a pytest fixture. This means that
most decorated functions can be handled (note: this still does not work
with certain pytest.parametrize'd functions).

Also added dependency handling for the class line, both in cut and in
splice (heuristically).

[File's changed]

	modified:   framework/bin/bugsinpy-cut
	modified:   framework/bin/bugsinpy-splice