Replaced the LCS java with a python version.
[Details]
The search for versions for this dataset used a longest common
subsequence algorithm that was written in java. This poses an additional
requirement on the user of this dataset to have java installed. To
remove this additional requirement, the algorithm was ported to a python
implementation. The new implementation was tested to ensure they perform
identically.

[Note]

[Bugs]

[File's changed]
	deleted:    lcs.class
	deleted:    lcs.java
	new file:   lcs.py