Added the pytest timeout to coverage collection.
[Details]
Due to some of the test cases in some projects getting stuck and never
terminating, the pytest-timeout plugin was added to the coverage
collection script, and used when the full test suite is run for
coverage. This allows individual test cases to be timed-out when running
longer than 5 minutes (value can be altered).

[File's changed]

	modified:   framework/bin/bugsinpy-coverage