Added pytest-timeout to coverage & fixed bugs.
[Details]
Added the pytest-timeout plugin when measuring coverage using pytest.
Some of the projects had test cases that would hang for unknown reasons
when running the coverage generation. To prevent this, the
pytest-timeout plugin kills test cases that take too long to run.

[Note]
Multiple versions of the pytest-timeout plugin are used for each
supported python version.

[Bugs]
- Fixed a bug in the coverage generation where unittest was attempting to
  be used for tox projects.
- Fixed a bug in the coverage where the coverage report was not being
  generated because of "missing source files"

[File's changed]
	modified:   framework/bin/bugsinpy-coverage