Changed the to-tcm to remove test dir names.
[Details]
In some projects, there is a difference in the way coverage.py displays
the names of test cases in various locations. The difference is in the
omission of the test directory name (usually "test" or "tests"). To
counteract this (and for test case name brevity) the bugsinpy-to-tcm
script now removes only the test directory name from the qualified test
case name in the TCM.

[File's changed]
	modified:   framework/bin/bugsinpy-to-tcm