Updated the LCS algorithm.
[Details]
Updated the LCS algorithm to consider each line separately. To keep the
efficiency, the algorithm checks whether it can efficiently compare
every line in the first input with every line in the second. If it can,
it will do so, otherwise it will only consider certain subsamples.

This update makes the LCS algorithm much more reliable.

[File's changed]
	modified:   lcs.py