Fixed bug in using poetry.
[Bugs]
The latest version of poetry (1.8.3) had a bug in it that caused it to
crash when using old versions of python (< 3.7). This was fixed in the
one of the latest commits for poetry (+-2 days ago). To apply this, the
commit version from github using the fix is used to install poetry.

[File's changed]
	modified:   bugsinpy-coverage