Fixed poetry virtual environment bug.
[Bugs]
It was assumed that poetry will always install the dependencies to the
already activated virtual environment 'env' in each project. However
poetry was creating a separate virtual environment inside the cache
directory in the users home. To solve this, a symlink of the env
directory was created called .venv, which poetry now recognizes as the
virtual environment for the project.

[File's changed]
	modified:   bugsinpy-coverage