Generalized the identify script.
[Details]
Made some improvements to the bugsinpy-identify script to make it usable
in a general situation. The script can now be called from a directory
that is not a checked out project. If this is done, the project and
version of the coverage file to identify must be given. The specific
coverage TCM file to use may also be specified instead of the default
coverage.tcm. These improvements are particularly useful when running
the identify script on an already created TCM file that has been moved.

[File's changed]
	modified:   bugsinpy-identify