Added method level collection for TCMs.
[Details]
Feature: Collection of method level information for the TCM spectra for
all BugsInPy python projects.
Description: By default, coverage.py does not collect method level
information when collecting the execution spectra. It only collects
the file and line number for the executed element. In order to add
method level information, this commit introduces changes that look at
the original project source code, and maps line numbers to methods. The
functionality is by default enabled for the bugsinpy-to-tcm script, and
can be run when generating a new TCM or on an already created one.

[Note]
The bugsinpy-identify script was modified to respect method level
information as well now.

[File's changed]
	modified:   bugsinpy-identify
	modified:   bugsinpy-to-tcm
	new file:   method_defs.py